+++ perf/read.ts
@@ -7,7 +7,6 @@ import { Quadstore } from '../lib/quadstore';
 import { disk, time } from './utils';
 import { waitForEvent } from '../lib/utils';
 import { DataFactory } from 'rdf-data-factory';
-import { newEngine } from 'quadstore-comunica';
 
 const dataFactory = new DataFactory();
 const qty = 200000;
@@ -40,7 +39,6 @@ disk(async (backend, checkDiskUsage) => {
   const store = new Quadstore({
     backend,
     dataFactory,
-    comunica: newEngine(),
   });
   await store.open();
   await doWrites(store);