+++ test/sparql/sparql.update.js
@@ -1,8 +1,0 @@
-
-module.exports = () => {
-
-  require('./sparql.update.deletedata')();
-  require('./sparql.update.insertdata')();
-  require('./sparql.update.insertdelete')();
-
-};